Distinguish the bye packet interpretation
In openconnect client the BYE packet indicates an explicit
user disconnect by sending 0x0b as payload. In anyconnect clients it
may indicate an intention to reconnect (e.g., because network was changed).
We introduce a check for 0x0b to identify the user disconnect and
add debugging output for other disconnect reasons.

Relates: #281

Signed-off-by: Nikos Mavrogiannopoulos <nmav@redhat.com>